+++ src/main/java/com/snow/popin/domain/mission/constant/UserMissionStatus.java
@@ -1,4 +1,4 @@
-package com.snow.popin.domain.mission.entity;
+package com.snow.popin.domain.mission.constant;
 
 public enum UserMissionStatus {
     PENDING, COMPLETED, FAIL
