+++ src/main/java/com/snow/popin/domain/mission/dto/response/MissionSetAdminDto.java
@@ -1,7 +1,7 @@
-package com.snow.popin.domain.mission.dto;
+package com.snow.popin.domain.mission.dto.response;
 
 import com.snow.popin.domain.mission.entity.MissionSet;
-import com.snow.popin.domain.mission.entity.MissionSetStatus;
+import com.snow.popin.domain.mission.constant.MissionSetStatus;
 import lombok.Builder;
 import lombok.Getter;
