+++ src/main/java/com/snow/popin/domain/mission/dto/response/MissionSummaryDto.java
@@ -1,6 +1,6 @@
-package com.snow.popin.domain.mission.dto;
+package com.snow.popin.domain.mission.dto.response;
 
-import com.snow.popin.domain.mission.entity.UserMissionStatus;
+import com.snow.popin.domain.mission.constant.UserMissionStatus;
 import lombok.Builder;
 import lombok.Getter;
