+++ src/main/java/com/snow/popin/domain/notification/dto/request/NotificationSettingRequestDto.java
@@ -1,4 +1,4 @@
-package com.snow.popin.domain.notification.dto;
+package com.snow.popin.domain.notification.dto.request;
 
 import lombok.Getter;
