+++ src/main/java/com/snow/popin/domain/reward/constant/UserRewardStatus.java
@@ -1,3 +1,3 @@
-package com.snow.popin.domain.reward.entity;
+package com.snow.popin.domain.reward.constant;
 
 public enum UserRewardStatus { ISSUED, REDEEMED, CANCELED }
