+++ src/main/java/com/snow/popin/domain/reward/repository/UserRewardRepository.java
@@ -1,7 +1,7 @@
 package com.snow.popin.domain.reward.repository;
 
 import com.snow.popin.domain.reward.entity.UserReward;
-import com.snow.popin.domain.reward.entity.UserRewardStatus;
+import com.snow.popin.domain.reward.constant.UserRewardStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.*;
